Infer default model_deployment in crfm-models server
