+++ src/helm/benchmark/run_spec_factory.py
@@ -1,11 +1,10 @@
 import dataclasses
-from typing import List, Optional
+from typing import List
 
 from helm.benchmark.adaptation.adapter_spec import ADAPT_GENERATION, ADAPT_MULTIPLE_CHOICE_JOINT
 from helm.benchmark.model_deployment_registry import (
-    ALL_MODEL_DEPLOYMENTS,
-    DEPLOYMENT_NAME_TO_MODEL_DEPLOYMENT,
     ModelDeployment,
+    get_default_model_deployment_for_model,
     get_model_deployment,
 )
 from helm.benchmark.model_metadata_registry import (
@@ -39,82 +38,7 @@ from helm.benchmark.run_expander import (
 )
 from helm.benchmark.run_spec import RunSpec, get_run_spec_function
 from helm.common.general import singleton
-from helm.common.hierarchical_logger import hlog
 from helm.common.object_spec import ObjectSpec
-
-
-def get_default_model_deployment_for_model(
-    model_name: str, warn_arg_deprecated: bool = False, ignore_deprecated: bool = False
-) -> Optional[str]:
-    """Returns a valid model deployment name corresponding to the given model arg.
-    This is used as a backwards compatibility layer for model names that are now moved to model deployments.
-    Example: "anthropic/claude-v1.3" => "anthropic/claude-v1.3"
-    Example: "meta/llama-7b" => "together/llama-7b"
-
-    The process to find a model deployment name is as follows:
-    1. If there is a model deployment with the same name as the model arg, use it.
-    2. If there is at least one deployment for the model, use the first one that is available.
-    3. If there are no deployments for the model, returns None.
-
-    This function will also try to find a model deployment name that is not deprecated.
-    If there are no non-deprecated deployments, it will return the first deployment (even if it's deprecated).
-    If ignore_deprecated is True, this function will return None if the model deployment is deprecated.
-
-    If warn_arg_deprecated is True, this function will print a warning if the model deployment name is not the same
-    as the model arg. This is to remind the user that the model name is deprecated and should be replaced with
-    the model deployment name (in their config).
-
-    Args:
-        model_arg: The model arg to convert to a model deployment name.
-        warn_arg_deprecated: Whether to print a warning if the model deployment name is not the same as the model arg.
-        ignore_deprecated: Whether to return None if the model deployment is deprecated.
-    """
-
-    # If there is a model deployment with the same name as the model arg, use it.
-    if model_name in DEPLOYMENT_NAME_TO_MODEL_DEPLOYMENT:
-        deployment: ModelDeployment = DEPLOYMENT_NAME_TO_MODEL_DEPLOYMENT[model_name]
-        if deployment.deprecated and ignore_deprecated:
-            if warn_arg_deprecated:
-                hlog(f"WARNING: Model deployment {model_name} is deprecated")
-            return None
-        return deployment.name
-
-    # If there is at least one deployment for the model, use the first one that is available.
-    available_deployments: List[ModelDeployment] = [
-        deployment for deployment in ALL_MODEL_DEPLOYMENTS if deployment.model_name == model_name
-    ]
-    if len(available_deployments) > 0:
-        available_deployment_names: List[str] = [deployment.name for deployment in available_deployments]
-        if warn_arg_deprecated:
-            hlog("WARNING: Model name is deprecated. Please use the model deployment name instead.")
-            hlog(f"Available model deployments for model {model_name}: {available_deployment_names}")
-
-        # Additionally, if there is a non-deprecated deployment, use it.
-        non_deprecated_deployments: List[ModelDeployment] = [
-            deployment for deployment in available_deployments if not deployment.deprecated
-        ]
-        if len(non_deprecated_deployments) > 0:
-            chosen_deployment = non_deprecated_deployments[0]
-        # There are no non-deprecated deployments, so there are two options:
-        # 1. If we can return an empty string, return it. (no model deployment is available)
-        # 2. If we can't return an empty string, return the first deployment (even if it's deprecated).
-        elif ignore_deprecated:
-            return None
-        else:
-            chosen_deployment = available_deployments[0]
-            if warn_arg_deprecated:
-                hlog(f"WARNING: All model deployments for model {model_name} are deprecated.")
-        if warn_arg_deprecated:
-            hlog(
-                f"Choosing {chosen_deployment.name} (the first one) as "
-                f"the default model deployment for model {model_name}"
-            )
-            hlog("If you want to use a different model deployment, please specify it explicitly.")
-        return chosen_deployment.name
-
-    # Some models are added but have no deployments yet.
-    # In this case, we return None.
-    return None
 
 
 def construct_run_specs(spec: ObjectSpec) -> List[RunSpec]:
